[Grupo 6] - Refatorado com custom hooks (#306)
* Refatorado codigo food e drinks

* Criado custom hooks
