+++ src/pages/Drinks.jsx
@@ -1,71 +1,29 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { Redirect, useLocation } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import { requestDrinks } from '../redux/actions/recipesActions';
-import { sendCocktailData } from '../redux/actions/cocktailsActions';
-import {
-  getDataByIngredient,
-  getDataByName,
-  getDataByFirstLetter,
-} from '../services/api';
+import { useSelector } from 'react-redux';
 import Header from '../components/Header';
 import SearchBar from '../components/SearchBar';
 import Card from '../components/Card';
+import useDrinks from '../hooks/useDrinks';
 
 export default function Drinks() {
-  const [data, setData] = useState([]);
-  const [inputValue, setInputValue] = useState('');
-  const [searched, setSearched] = useState(false);
-  const dispatch = useDispatch();
   const [showSearchBar, setShowSearchBar] = useState(false);
-  const [filterIngredient, setFilterIngredient] = useState(false);
-  const [filterName, setFilterName] = useState(false);
-  const [filterFirstLetter, setFilterFirstLetter] = useState(false);
-
-  const drinks = useSelector((state) => state.recipes.recipes);
-
-  useEffect(() => {
-    dispatch(requestDrinks());
-  }, [dispatch]);
-
   const location = useLocation();
   const currentPage = location.pathname;
-
-  const qtd = 12;
-  console.log(drinks);
+  const {
+    setInputValue,
+    setFilterIngredient,
+    setFilterName,
+    setFilterFirstLetter,
+    setSearched,
+    inputValue,
+  } = useDrinks();
+  const drinks = useSelector((state) => state.recipes.recipes);
 
   const handleClick = async (e) => {
     e.preventDefault();
-    dispatch(sendCocktailData({ data }));
     setSearched(true);
   };
-
-  useEffect(() => {
-    if (searched) {
-      if (filterIngredient && !filterName && !filterFirstLetter) {
-        getDataByIngredient(inputValue)
-          .then((response) => setData(response.drinks));
-      } else if (filterName && !filterIngredient && !filterFirstLetter) {
-        getDataByName(inputValue)
-          .then((response) => setData(response.drinks));
-      } else if (filterFirstLetter && !filterIngredient && !filterName) {
-        getDataByFirstLetter(inputValue)
-          .then((response) => setData(response.drinks));
-      }
-    }
-  }, [
-    filterIngredient,
-    filterName,
-    filterFirstLetter,
-    currentPage,
-    inputValue,
-    searched]);
-
-  if (data === null) {
-    // eslint-disable-next-line no-alert
-    alert('Sinto muito, não encontramos nenhuma receita para esses filtros.');
-    return data;
-  }
 
   return (
     <section>
@@ -83,23 +41,12 @@ export default function Drinks() {
           inputValue={ inputValue }
         />
       ) : null}
-      {data.length === 1 ? (
-        data.map(({ idDrink }) => (
+      {drinks.length === 1 ? (
+        drinks.map(({ idDrink }) => (
           <Redirect key={ idDrink } to={ `${currentPage}/${idDrink}` } />
         ))
       ) : null}
-      {searched ? (
-        data
-          .slice(0, qtd)
-          .map(({ strDrink, idDrink, strDrinkThumb }, index) => (
-            <Card
-              key={ idDrink }
-              thumb={ strDrinkThumb }
-              title={ strDrink }
-              index={ index }
-            />
-          ))
-      ) : (
+      {
         drinks.map((drink, index) => (
           <Card
             key={ drink.idDrink }
@@ -108,7 +55,7 @@ export default function Drinks() {
             index={ index }
           />
         ))
-      )}
+      }
     </section>
   );
 }
