--- src/redux/actions/cocktailsActions.js
@@ -1,5 +1,0 @@
-export const SEND_COCKTAIL_DATA = 'SEND_COCKTAIL_DATA';
-
-export const sendCocktailData = (info) => ({
-  type: SEND_COCKTAIL_DATA, info,
-});
