--- src/redux/actions/recipeActions.js
@@ -1,5 +1,0 @@
-export const SEND_RECIPE_DATA = 'SEND_RECIPE_DATA';
-
-export const sendRecipeData = (info) => ({
-  type: SEND_RECIPE_DATA, info,
-});
