+++ src/redux/actions/recipesActions.js
@@ -9,19 +9,19 @@ export const loadingRecipes = () => ({
 });
 
 export const getRecipes = (recipes) => ({
-  type: GET_RECIPES, recipes,
+  type: GET_RECIPES, recipes: recipes.slice(0, RECIPES_QUANTITY),
 });
 
 export const requestFoods = () => async (dispatch) => {
   dispatch(loadingRecipes());
   const response = await fetch(FOODS_URL);
   const { meals } = await response.json();
-  dispatch(getRecipes(meals.slice(0, RECIPES_QUANTITY)));
+  dispatch(getRecipes(meals));
 };
 
 export const requestDrinks = () => async (dispatch) => {
   dispatch(loadingRecipes());
   const response = await fetch(DRINKS_URL);
   const { drinks } = await response.json();
-  dispatch(getRecipes(drinks.slice(0, RECIPES_QUANTITY)));
+  dispatch(getRecipes(drinks));
 };
