+++ src/redux/reducers/index.js
@@ -1,6 +1,5 @@
 import { combineReducers } from 'redux';
 import user from './user';
 import recipes from './recipes';
-import cocktail from './cocktail';
 
-export default combineReducers({ user, recipes, cocktail });
+export default combineReducers({ user, recipes });
